Enemy Spawn System
Added a new empty enemies object to hold all enemy objects with a script that spawns an enemy every 5 seconds, the time can be customized and even randomized later on if we want. Spawned enemies will have a 75/25 ratio of basic enemies to fast enemies. This function should still work with any level as long as there is a NavMesh, we will just need to change the radius enemies spawn in for bigger levels.